+++ orm/src/main/java/org/hibernate/tool/internal/export/pojo/ImportContextImpl.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbm2x.pojo;
+package org.hibernate.tool.internal.export.pojo;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -7,7 +7,6 @@ import java.util.Set;
 import java.util.TreeSet;
 
 import org.hibernate.internal.util.StringHelper;
-import org.hibernate.tool.internal.export.pojo.ImportContext;
 
 public class ImportContextImpl implements ImportContext {
